Added the font-awesome library. This allows me to implement icons into the project.
Created the AddBackgroundImageModalContent component. This is currently just a very basic component, but the content in this component is appearing in the modal that appears when the "Add background image" button is pressed.

Created the CanvassTopNavBar compnent. The only button that is currently funciton in this component is the "Add background image" component. Will need to add functionality to the other buttons as well.

Added the CanvassTopNavBar component into the LandscapeCanvass component.